Do not treat environment filter failure as terminal
We only enable perfmap-based symbolization for a process if it has `__PERFORATOR_ENABLE_PERFMAP` environment variable set. Moreover, this check is evaluated only once, when we first discover this process. If check fails, we treat such a process as permanently ineligible.

This logic turned out to be incorrect:
* Sometimes querying environment variables for a newly created process returns empty map, and filter runs on that map.
* When using execve syscall, process environment can change in arbitrary ways. If a JVM process is started indirectly using execve-based launcher, filter runs on a launcher environment instead of actual JVM environment.

In both cases we end up skipping processes, which decreases quality.

As a fix, perfmap registry now subscribes to Rescan events (they are currently emitted for a process when the set of its mappings changes) and uses them as additional chance to apply environment filter.
This should fix second issue and help with fixing first one.
commit_hash:7bff58029037efe3841e9692910a94d131ea1581